Developed a smarter display of temperatures
The build and program page now display live readings based of the last button pressed and they give correct error messages, There is still a need to develop a display for the chance that more than one sensor is missing. an idea is to pass a list of sensor positions rather than a string message.