build button/program buttons added
For the momment the test button on the build test hasnt been fully tested and can be highly improved, It does run and doesnt crash,

Program on the other hand isnt fully functional there is a loop that the eeprom program gets stuck waiting for the output.  There needs to be better implementation in the way that data is being requested when reading and writing.